// src/types.ts
export type UserRole = 
  | 'Super Admin' 
  | 'Admin'
  | 'Manager Resepsionis' 
  | 'Resepsionis' 
  | 'Manager QC' 
  | 'Quality Control' 
  | 'Manager Teknisi' 
  | 'Teknisi' 
  | 'Manager Koperasi' 
  | 'Petugas Koperasi'
  | string;

export interface User {
  id: string;
  username: string;
  fullName: string;
  role: UserRole;
  password?: string;
  department?: string;
  supervisorId?: string | null;
  assignedBuilding: string;
  phone: string;
  status: string;
  email?: string;
  isOwner?: boolean;
}